requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
sentence-transformers==2.2.2
chromadb==0.4.22
huggingface-hub==0.16.4
langchain==0.0.350
langchain-community==0.0.10
pandas==2.1.3
numpy==1.24.3
mecab-python3==1.0.8
jinja2==3.1.2
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
httpx==0.25.1
tenacity==8.2.3
loguru==0.7.2